requirements.txt
pytorch-lightning =1.6.3=pypi_0
pytorch=1.11.0=py3.9_cuda11.3_cudnn8.2.0_0
torchvision=0.12.0=py39_cu113
torchmetrics=0.8.2=pypi_0
